+++ src/components/landing/LandingSection2.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 import React from 'react';
 import styled from 'styled-components';
-import { Depth } from '../post/Modal';
+import { Depth } from '../common/post/Modal';
 
 interface DataType extends Depth {
 	author: string;
